% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{available_systems}
\alias{available_systems}
\title{List supported historical bike-share systems}
\usage{
available_systems()
}
\value{
A tibble with system identifiers, names, cities, and official data
pages.
}
\description{
List supported historical bike-share systems
}
