% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-validate.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Bike-rental data dictionary}
\usage{
data_dictionary()
}
\value{
A tibble describing the processed dataset variables.
}
\description{
Bike-rental data dictionary
}
