% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{print.bikerental_system_info}
\alias{print.bikerental_system_info}
\title{Print current Capital Bikeshare system information}
\usage{
\method{print}{bikerental_system_info}(x, ...)
}
\arguments{
\item{x}{A \code{bikerental_system_info} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print current Capital Bikeshare system information
}
