% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{trip_data_dictionary}
\alias{trip_data_dictionary}
\title{Standardized trip-data dictionary}
\usage{
trip_data_dictionary()
}
\value{
A tibble describing every field returned by
\code{\link[=standardize_trips]{standardize_trips()}} and \code{\link[=load_trip_data]{load_trip_data()}}.
}
\description{
Standardized trip-data dictionary
}
